removed dependency from template (in Library/math)

+++ math/modinv.hpp
@@ -1,9 +1,8 @@
-// ※テンプレート依存
 // find y(0 <= y < mod) where modx + ay = 1
-ll modinv(ll a, ll mod){
-    ll b = mod, y = 1, x = 0;
+long long modinv(long long a, long long mod){
+    long long b = mod, y = 1, x = 0;
     while(b){
-        ll t = a / b;
+        long long t = a / b;
         a -= t * b; swap(a, b);
         y -= t * x; swap(y, x);
     }
